requirements.txt
python-dotenv
streamlit
langchain-groq==0.1.9
langgraph==0.2.1
openbb==4.3.1
pandas_ta==0.3.14b
duckduckgo_search
numpy==1.26.4
flask
flask_cors
yfinance
finnhub-python
requests
pydantic>=2.0.0,<3.0.0
google-generativeai
